+++ modules/c++/net/source/CurlHandle.cpp
@@ -98,10 +98,10 @@ void CurlHandle::verify(CURLcode code, const std::string& prefix)
     }
 }
 
-int CurlHandle::writeCallback(char* data,
-                              size_t size,
-                              size_t nmemb,
-                              std::string* writeData)
+size_t CurlHandle::writeCallback(char* data,
+                                 size_t size,
+                                 size_t nmemb,
+                                 std::string* writeData)
 {
     if (writeData == NULL)
     {
